+++ src/video_server/file.clj
@@ -178,14 +178,14 @@
   (some? (exts (file-ext file))))
 
 (defn dir-filter
-  ^FilenameFilter []
+  ^java.io.FilenameFilter []
   (reify FilenameFilter
     (accept [_ dir name] (and (not (hidden? name))
                               (dir? (io/file dir name))))))
 
 (defn ext-filter
   "Returns a filename filter that matches the extensions."
-  ^FilenameFilter [exts]
+  ^java.io.FilenameFilter [exts]
   (reify FilenameFilter
     (accept [_ dir name] (and (not (hidden? name))
                               (file-with-ext? name exts)))))
@@ -193,7 +193,7 @@
 (defn title-filter
   "Returns a filename filter that matches the video title and
   extensions."
-  ^FilenameFilter [title exts]
+  ^java.io.FilenameFilter [title exts]
   (let [title (str/lower-case (clean-title title))]
     (reify FilenameFilter
       (accept [this dir name]
@@ -223,18 +223,18 @@
 
 (defn movie-filter
   "A filter for listing movie files."
-  ^FilenameFilter []
+  ^java.io.FilenameFilter []
   (ext-filter movie-exts))
 
 (defn subtitle-filter
   "A filter for listing subtitle files. If a title is specified, the
   filter will only match files related to that title."
-  (^FilenameFilter [] (ext-filter subtitle-exts))
-  (^FilenameFilter [title] (title-filter title subtitle-exts)))
+  (^java.io.FilenameFilter [] (ext-filter subtitle-exts))
+  (^java.io.FilenameFilter [title] (title-filter title subtitle-exts)))
 
 (defn image-filter
   "A filter for listing image files. If a title is specified, the
   filter will only match files related to that title."
-  (^FilenameFilter [] (ext-filter image-exts))
-  (^FilenameFilter [title] (title-filter title image-exts)))
+  (^java.io.FilenameFilter [] (ext-filter image-exts))
+  (^java.io.FilenameFilter [title] (title-filter title image-exts)))